pyproject.toml
[project]
name = "outerworldswiki"
version = "0.1.0"
description = ""
requires-python = "~=3.10"
readme = "README.md"
dependencies = [
    "dotenv>=0.9.9",
    "openai>=1.47.0,<2",
    "pandas>=2.2.3",
    "pillow>=11.1.0",
    "requests>=2.32.3",
]

[tool.uv]
package = false

[tool.ruff]
line-length = 100
target-version = "py310"

[tool.ruff.lint]
select = ["E4", "E7", "E9", "F"]
ignore = ["F401"]

[tool.ruff.format]
quote-style = "double"
